fix(ci): unblock PR 130 checks

+++ packages/desktop/src-tauri/src/cc_sdk/transport/subprocess.rs
@@ -1297,7 +1297,7 @@ fn apply_process_user_inner(cmd: &mut Command, user: &str) -> Result<()> {
         }
 
         let pwd = unsafe { pwd.assume_init() };
-        Ok((pwd.pw_uid as u32, pwd.pw_gid as u32))
+        Ok((pwd.pw_uid, pwd.pw_gid))
     }
 
     fn lookup_by_uid(uid: u32) -> Result<(u32, u32)> {
@@ -1328,7 +1328,7 @@ fn apply_process_user_inner(cmd: &mut Command, user: &str) -> Result<()> {
         }
 
         let pwd = unsafe { pwd.assume_init() };
-        Ok((pwd.pw_uid as u32, pwd.pw_gid as u32))
+        Ok((pwd.pw_uid, pwd.pw_gid))
     }
 
     let (uid, gid) = match user.parse::<u32>() {
